requirements.txt
cloudpickle==2.2.1
Cython==0.29.35
Flask==2.3.2
gcsfs==2023.6.0
google-api-python-client==1.8.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
imageio==2.31.1
iopath==0.1.9
matplotlib==3.7.1
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
pandas==1.5.3
Pillow==9.5.0
plotly==5.15.0
protobuf==3.20.3
pyarrow==12.0.1
torch==1.13.1
torchvision==0.14.1
Werkzeug==2.3.6
gunicorn==20.1.0
opencv-python-headless==4.8.0.74